requirements.txt
pdfplumber>=0.11.0
pypdf>=4.0.0
python-dotenv>=1.0.0
requests>=2.31.0
beautifulsoup4>=4.12.0
composio-core>=0.7.0
